Frequency as float, new plotting

+++ oscillator.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 class Oscillator(ABC):
     def __init__(
         self,
-        freq: int = 440,
+        freq: float = 440.0,
         phase: int = 0,
         amp: float = 1.0,
         sample_rate: int = 44100,
